+++ src/features/reader/overlay/navigation/desktop/components/ReaderNavBarDesktopChapterNavigation.tsx
@@ -15,7 +15,6 @@ import Popover from '@mui/material/Popover';
 import { bindPopover, bindTrigger, usePopupState } from 'material-ui-popup-state/hooks';
 import FormControl from '@mui/material/FormControl';
 import InputLabel from '@mui/material/InputLabel';
-import { Link } from 'react-router-dom';
 import { useLingui } from '@lingui/react/macro';
 import { Select } from '@/base/components/inputs/Select.tsx';
 import { ReaderChapterList } from '@/features/reader/overlay/navigation/components/ReaderChapterList.tsx';
@@ -26,6 +25,7 @@ import { ReaderControls } from '@/features/reader/services/ReaderControls.ts';
 import type { ReaderStateChapters } from '@/features/reader/Reader.types.ts';
 import { withPropsFrom } from '@/base/hoc/withPropsFrom.tsx';
 import type { ChapterIdInfo } from '@/features/chapter/Chapter.types.ts';
+import { useTheme } from '@mui/material/styles';
 
 const BaseReaderNavBarDesktopChapterNavigation = ({
     currentChapterId,
@@ -43,6 +43,7 @@ const BaseReaderNavBarDesktopChapterNavigation = ({
         readerThemeDirection: ReturnType<typeof ReaderService.useGetThemeDirection>;
     }) => {
     const { t } = useLingui();
+    const theme = useTheme();
 
     const popupState = usePopupState({ variant: 'popover', popupId: 'reader-nav-bar-desktop-chapter-list' });
 
@@ -60,7 +61,7 @@ const BaseReaderNavBarDesktopChapterNavigation = ({
                 }}
                 disabled={getOptionForDirection(!previousChapter, !nextChapter, readerThemeDirection)}
             />
-            <FormControl sx={{ flexBasis: '70%', flexGrow: 0, flexShrink: 0 }}>
+            <FormControl sx={{ flexBasis: '70%', flexGrow: 0, flexShrink: 0 }} dir={theme.direction}>
                 <InputLabel id="reader-nav-bar-desktop-chapter-select">{t`Chapter`}</InputLabel>
                 <Select
                     {...bindTrigger(popupState)}
@@ -78,7 +79,6 @@ const BaseReaderNavBarDesktopChapterNavigation = ({
                 </Select>
             </FormControl>
             <ReaderNavBarDesktopNextPreviousButton
-                component={Link}
                 type="next"
                 title={getOptionForDirection(t`Next chapter`, t`Previous chapter`, readerThemeDirection)}
                 onClick={() => {