+++ src/features/reader/settings/general/components/ReaderSettingBackgroundColor.tsx
@@ -12,6 +12,8 @@ import type { ValueToDisplayData } from '@/base/Base.types.ts';
 import { ButtonSelectInput } from '@/base/components/inputs/ButtonSelectInput.tsx';
 import type { IReaderSettingsWithDefaultFlag } from '@/features/reader/Reader.types.ts';
 import { ReaderBackgroundColor } from '@/features/reader/Reader.types.ts';
+import type { ReaderService } from '@/features/reader/services/ReaderService.ts';
+import { CheckboxInput } from '@/base/components/inputs/CheckboxInput';
 
 const VALUE_TO_DISPLAY_DATA: ValueToDisplayData<ReaderBackgroundColor> = {
     [ReaderBackgroundColor.THEME]: {
@@ -30,6 +32,10 @@ const VALUE_TO_DISPLAY_DATA: ValueToDisplayData<ReaderBackgroundColor> = {
         title: msg`White`,
         icon: null,
     },
+    [ReaderBackgroundColor.AUTO]: {
+        title: msg`Auto`,
+        icon: null,
+    },
 };
 
 const READER_BACKGROUND_COLOR_VALUES = Object.values(ReaderBackgroundColor).filter(
@@ -38,19 +44,29 @@ const READER_BACKGROUND_COLOR_VALUES = Object.values(ReaderBackgroundColor).filt
 
 export const ReaderSettingBackgroundColor = ({
     backgroundColor,
+    useAutoBackgroundColorContinuousMode,
     updateSetting,
-}: Pick<IReaderSettingsWithDefaultFlag, 'backgroundColor'> & {
-    updateSetting: (color: ReaderBackgroundColor) => void;
+}: Pick<IReaderSettingsWithDefaultFlag, 'backgroundColor' | 'useAutoBackgroundColorContinuousMode'> & {
+    updateSetting: (...args: Parameters<typeof ReaderService.updateSetting>) => void;
 }) => {
     const { t } = useLingui();
 
     return (
-        <ButtonSelectInput
-            label={t`Background color`}
-            value={backgroundColor}
-            values={READER_BACKGROUND_COLOR_VALUES}
-            setValue={updateSetting}
-            valueToDisplayData={VALUE_TO_DISPLAY_DATA}
-        />
+        <>
+            <ButtonSelectInput
+                label={t`Background color`}
+                value={backgroundColor}
+                values={READER_BACKGROUND_COLOR_VALUES}
+                setValue={(value) => updateSetting('backgroundColor', value)}
+                valueToDisplayData={VALUE_TO_DISPLAY_DATA}
+            />
+            {backgroundColor === ReaderBackgroundColor.AUTO && (
+                <CheckboxInput
+                    label={t`Enable auto background color in continuous reading mode`}
+                    checked={useAutoBackgroundColorContinuousMode}
+                    onChange={(_, checked) => updateSetting('useAutoBackgroundColorContinuousMode', checked)}
+                />
+            )}
+        </>
     );
 };